+++ history.md
@@ -1,8 +1,14 @@
 
-2014-08-03/01
+2014-08-03
 ===
 
+### 更正插件的中文名称
+
 * 修正 MerchantEx、PetJournalEnhanced 的中文名称
+
+### 聊天过滤
+
+* 经过一段时间的实测发现 ChannelClean 的拦截效果不如 ChatFilter ；因此移除ChannelClean，重新换回了 ChatFilter，同时优化了 ChatFilter 的占用，并处理掉了自动乱加好友的问题
 
 
 2014-07-30/01